backend/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
langchain-chroma>=0.0.1
langgraph>=0.0.20
chromadb>=0.4.22
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.8.0
tiktoken>=0.5.0 
requests>=2.31.0

numpy
fastapi>=0.100.0

# LLM Provider packages (at least one is required)
# Comment out any provider you don't need
ollama>=0.4.0
openai>=1.7.0
anthropic>=0.5.0
google-generativeai>=0.3.0
langchain_xai
